+++ gsa/src/web/pages/extras/cvsscalculatorpage.js
@@ -24,7 +24,12 @@ import _ from 'gmp/locale';
 import {KeyCode} from 'gmp/utils/event';
 import {isDefined} from 'gmp/utils/identity';
 
-import {parseCvssBaseVector, parseCvssBaseFromVector} from 'gmp/parser';
+import {
+  parseCvssV2BaseVector,
+  parseCvssV3BaseVector,
+  parseCvssV2BaseFromVector,
+  parseCvssV3BaseFromVector,
+} from 'gmp/parser/cvss';
 
 import SeverityBar from 'web/components/bar/severitybar';
 
@@ -36,11 +41,11 @@ import CvssIcon from 'web/components/icon/cvssicon';
 import ManualIcon from 'web/components/icon/manualicon';
 
 import Layout from 'web/components/layout/layout';
+import Divider from 'web/components/layout/divider';
 
 import Section from 'web/components/section/section';
 
 import useUserSessionTimeout from 'web/utils/useUserSessionTimeout';
-import useGmp from 'web/utils/useGmp';
 
 const StyledTextField = styled(TextField)`
   width: 180px;
@@ -55,8 +60,9 @@ const ToolBarIcons = () => (
   />
 );
 
-const CvssCalculator = props => {
-  const gmp = useGmp();
+/* CVSS v2 .... */
+
+const CvssV2Calculator = props => {
   const [, renewSession] = useUserSessionTimeout();
 
   const [state, setState] = useState({
@@ -80,9 +86,10 @@ const CvssCalculator = props => {
       isDefined(location.query.cvssVector)
     ) {
       const {cvssVector} = location.query;
-      setState(vals => ({...vals, cvssVector, userVector: cvssVector}));
-      calculateScore(cvssVector);
-      handleVectorChange();
+      if (!cvssVector.includes('CVSS:3')) {
+        setState(vals => ({...vals, cvssVector, userVector: cvssVector}));
+        handleVectorChange();
+      }
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
@@ -102,7 +109,7 @@ const CvssCalculator = props => {
       availabilityImpact,
     } = state;
 
-    const cvssVector = parseCvssBaseVector({
+    const [cvssVector, cvssScore] = parseCvssV2BaseVector({
       accessComplexity,
       accessVector,
       authentication,
@@ -117,24 +124,12 @@ const CvssCalculator = props => {
       ...newVector,
       cvssVector,
       userVector: cvssVector,
+      cvssScore: cvssScore,
     }));
-
-    calculateScore(cvssVector);
-  };
-
-  const calculateScore = cvssVector => {
-    gmp.cvsscalculator.calculateScoreFromVector(cvssVector).then(response => {
-      const {data: cvssScore} = response;
-      setState(vals => ({...vals, cvssScore}));
-    });
-  };
-
-  const handleInteraction = () => {
-    renewSession();
   };
 
   const handleMetricsChange = (value, name) => {
-    handleInteraction();
+    renewSession();
 
     calculateVector({[name]: value});
   };
@@ -146,30 +141,39 @@ const CvssCalculator = props => {
   const handleVectorChange = () => {
     const {userVector} = state;
 
-    handleInteraction();
+    renewSession();
 
-    const cvssValues = parseCvssBaseFromVector(userVector);
     const {
       accessVector,
       accessComplexity,
-      confidentialityImpact,
       authentication,
+      confidentialityImpact,
       integrityImpact,
       availabilityImpact,
-    } = cvssValues;
+      cvssScore,
+    } = parseCvssV2BaseFromVector(userVector);
 
     if (
       isDefined(accessVector) &&
       isDefined(accessComplexity) &&
-      isDefined(confidentialityImpact) &&
       isDefined(authentication) &&
+      isDefined(confidentialityImpact) &&
       isDefined(integrityImpact) &&
-      isDefined(availabilityImpact)
+      isDefined(availabilityImpact) &&
+      isDefined(cvssScore)
     ) {
       /* only override cvss values and vector if user vector has valid input */
-
-      setState(vals => ({...vals, ...cvssValues, cvssVector: userVector}));
-      calculateScore(userVector);
+      setState(vals => ({
+        ...vals,
+        accessVector,
+        accessComplexity,
+        authentication,
+        confidentialityImpact,
+        integrityImpact,
+        availabilityImpact,
+        cvssVector: userVector,
+        cvssScore,
+      }));
     }
   };
 
@@ -181,25 +185,23 @@ const CvssCalculator = props => {
   };
 
   const {
-    accessComplexity,
     accessVector,
+    accessComplexity,
     authentication,
-    availabilityImpact,
     confidentialityImpact,
+    availabilityImpact,
+    integrityImpact,
     userVector,
     cvssScore,
     cvssVector,
-    integrityImpact,
   } = state;
 
   return (
-    <Layout flex="column">
-      <ToolBarIcons />
+    <Layout flex="column" grow>
       <Section
         img={<CvssIcon size="large" />}
         title={_('CVSSv2 Base Score Calculator')}
       />
-
       <h3>{_('From Metrics')}:</h3>
       <FormGroup title={_('Access Vector')}>
         <Select
@@ -354,6 +356,364 @@ const CvssCalculator = props => {
     </Layout>
   );
 };
+
+/* CVSS v3 .... */
+
+const CvssV3Calculator = props => {
+  const [, renewSession] = useUserSessionTimeout();
+
+  const [state, setState] = useState({
+    attackVector: 'NETWORK',
+    attackComplexity: 'LOW',
+    privilegesRequired: 'NONE',
+    userInteraction: 'NONE',
+    scope: 'UNCHANGED',
+    authentication: 'NONE',
+    integrityImpact: 'NONE',
+    availabilityImpact: 'NONE',
+    cvssVector: 'CVSS:3.1/AV:N/AC:L/PR:N/UI:N/S:U/C:N/I:N/A:N',
+    userVector: 'CVSS:3.1/AV:N/AC:L/PR:N/UI:N/S:U/C:N/I:N/A:N',
+    cvssScore: 0,
+  });
+
+  useEffect(() => {
+    const {location} = props;
+
+    if (
+      isDefined(location) &&
+      isDefined(location.query) &&
+      isDefined(location.query.cvssVector)
+    ) {
+      const {cvssVector} = location.query;
+      if (cvssVector.includes('CVSS:3')) {
+        setState(vals => ({...vals, cvssVector, userVector: cvssVector}));
+        handleVectorChange();
+      }
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
+
+  useEffect(() => {
+    handleVectorChange();
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [state.userVector]);
+
+  const calculateVector = newVector => {
+    const {
+      attackVector,
+      attackComplexity,
+      privilegesRequired,
+      userInteraction,
+      scope,
+      confidentialityImpact,
+      integrityImpact,
+      availabilityImpact,
+    } = state;
+
+    const [cvssVector, cvssScore] = parseCvssV3BaseVector({
+      attackVector,
+      attackComplexity,
+      privilegesRequired,
+      userInteraction,
+      scope,
+      confidentialityImpact,
+      integrityImpact,
+      availabilityImpact,
+      ...newVector,
+    });
+
+    setState(vals => ({
+      ...vals,
+      ...newVector,
+      cvssVector,
+      userVector: cvssVector,
+      cvssScore: cvssScore,
+    }));
+  };
+
+  const handleMetricsChange = (value, name) => {
+    renewSession();
+
+    calculateVector({[name]: value});
+  };
+
+  const handleInputChange = (value, name) => {
+    setState(vals => ({...vals, [name]: value}));
+  };
+
+  const handleVectorChange = () => {
+    const {userVector} = state;
+
+    renewSession();
+
+    const {
+      attackVector,
+      attackComplexity,
+      privilegesRequired,
+      userInteraction,
+      scope,
+      confidentialityImpact,
+      integrityImpact,
+      availabilityImpact,
+      cvssScore,
+    } = parseCvssV3BaseFromVector(userVector);
+
+    if (
+      isDefined(attackVector) &&
+      isDefined(attackComplexity) &&
+      isDefined(privilegesRequired) &&
+      isDefined(userInteraction) &&
+      isDefined(scope) &&
+      isDefined(confidentialityImpact) &&
+      isDefined(integrityImpact) &&
+      isDefined(availabilityImpact) &&
+      isDefined(cvssScore)
+    ) {
+      /* only override cvss values and vector if user vector has valid input */
+
+      setState(vals => ({
+        ...vals,
+        attackVector,
+        attackComplexity,
+        privilegesRequired,
+        userInteraction,
+        scope,
+        confidentialityImpact,
+        integrityImpact,
+        availabilityImpact,
+        cvssVector: userVector,
+        cvssScore,
+      }));
+    }
+  };
+
+  const handleKeyDown = event => {
+    const key_code = event.keyCode;
+    if (key_code === KeyCode.ENTER) {
+      handleVectorChange();
+    }
+  };
+
+  const {
+    attackVector,
+    attackComplexity,
+    privilegesRequired,
+    userInteraction,
+    scope,
+    availabilityImpact,
+    confidentialityImpact,
+    userVector,
+    cvssScore,
+    cvssVector,
+    integrityImpact,
+  } = state;
+
+  return (
+    <Layout flex="column" grow>
+      <Section
+        img={<CvssIcon size="large" />}
+        title={_('CVSSv3 Base Score Calculator')}
+      />
+      <h3>{_('From Metrics')}:</h3>
+      <FormGroup title={_('Attack Vector')}>
+        <Select
+          items={[
+            {
+              value: 'LOCAL',
+              label: _('Local'),
+            },
+            {
+              value: 'ADJACENT_NETWORK',
+              label: _('Adjacent'),
+            },
+            {
+              value: 'NETWORK',
+              label: _('Network'),
+            },
+            {
+              value: 'PHYSICAL',
+              label: _('Physical'),
+            },
+          ]}
+          name="attackVector"
+          value={attackVector}
+          menuPosition="adjust"
+          onChange={handleMetricsChange}
+        />
+      </FormGroup>
+      <FormGroup title={_('Attack Complexity')}>
+        <Select
+          items={[
+            {
+              value: 'LOW',
+              label: _('Low'),
+            },
+            {
+              value: 'HIGH',
+              label: _('High'),
+            },
+          ]}
+          name="attackComplexity"
+          value={attackComplexity}
+          menuPosition="adjust"
+          onChange={handleMetricsChange}
+        />
+      </FormGroup>
+      <FormGroup title={_('Privileges Required')}>
+        <Select
+          items={[
+            {
+              value: 'NONE',
+              label: _('None'),
+            },
+            {
+              value: 'LOW',
+              label: _('Low'),
+            },
+            {
+              value: 'HIGH',
+              label: _('High'),
+            },
+          ]}
+          name="privilegesRequired"
+          value={privilegesRequired}
+          menuPosition="adjust"
+          onChange={handleMetricsChange}
+        />
+      </FormGroup>
+      <FormGroup title={_('User Interaction')}>
+        <Select
+          items={[
+            {
+              value: 'NONE',
+              label: _('None'),
+            },
+            {
+              value: 'REQUIRED',
+              label: _('Required'),
+            },
+          ]}
+          name="userInteraction"
+          value={userInteraction}
+          menuPosition="adjust"
+          onChange={handleMetricsChange}
+        />
+      </FormGroup>
+      <FormGroup title={_('Scope')}>
+        <Select
+          items={[
+            {
+              value: 'UNCHANGED',
+              label: _('Unchanged'),
+            },
+            {
+              value: 'CHANGED',
+              label: _('Changed'),
+            },
+          ]}
+          name="scope"
+          value={scope}
+          menuPosition="adjust"
+          onChange={handleMetricsChange}
+        />
+      </FormGroup>
+      <FormGroup title={_('Confidentiality')}>
+        <Select
+          items={[
+            {
+              value: 'NONE',
+              label: _('None'),
+            },
+            {
+              value: 'LOW',
+              label: _('Low'),
+            },
+            {
+              value: 'HIGH',
+              label: _('High'),
+            },
+          ]}
+          name="confidentialityImpact"
+          value={confidentialityImpact}
+          onChange={handleMetricsChange}
+        />
+      </FormGroup>
+      <FormGroup title={_('Integrity')}>
+        <Select
+          items={[
+            {
+              value: 'NONE',
+              label: _('None'),
+            },
+            {
+              value: 'LOW',
+              label: _('Low'),
+            },
+            {
+              value: 'HIGH',
+              label: _('High'),
+            },
+          ]}
+          name="integrityImpact"
+          value={integrityImpact}
+          menuPosition="adjust"
+          onChange={handleMetricsChange}
+        />
+      </FormGroup>
+      <FormGroup title={_('Availability')}>
+        <Select
+          items={[
+            {
+              value: 'NONE',
+              label: _('None'),
+            },
+            {
+              value: 'LOW',
+              label: _('Low'),
+            },
+            {
+              value: 'HIGH',
+              label: _('High'),
+            },
+          ]}
+          name="availabilityImpact"
+          value={availabilityImpact}
+          menuPosition="adjust"
+          onChange={handleMetricsChange}
+        />
+      </FormGroup>
+
+      <h3>{_('From Vector')}:</h3>
+      <FormGroup title={_('CVSS v3.1 Vector')}>
+        <StyledTextField
+          name="userVector"
+          value={userVector}
+          onChange={handleInputChange}
+          onBlur={handleVectorChange}
+          onKeyDown={handleKeyDown}
+        />
+      </FormGroup>
+
+      <h3>{_('Results')}:</h3>
+      <FormGroup title={_('CVSS Base Vector')}>
+        <span>{cvssVector}</span>
+      </FormGroup>
+      <FormGroup title={_('Severity')}>
+        <SeverityBar severity={cvssScore} />
+      </FormGroup>
+    </Layout>
+  );
+};
+
+const CvssCalculator = props => (
+  <Layout flex="column">
+    <ToolBarIcons />
+    <Divider margin="20px" flex="row" align={['center', 'start']} grow>
+      <CvssV2Calculator {...props} />
+      <CvssV3Calculator {...props} />
+    </Divider>
+  </Layout>
+);
 
 export default CvssCalculator;